surrender, render quality
